ViewModel 분리
